requirements.txt
alembic==1.8.1
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
cachelib==0.9.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
filelock==3.7.1
Flask==2.1.3
Flask-Migrate==2.6.0
Flask-Script==2.0.5
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
h11==0.13.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
numpy==1.23.1
outcome==1.2.0
pandas==1.4.3
parsel==1.6.0
Protego==0.2.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.1
scrapy-selenium==0.0.7
selenium==4.3.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
tldextract==3.3.1
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
typing_extensions==4.3.0
urllib3==1.26.10
w3lib==1.22.0
Werkzeug==2.2.0
wsproto==1.1.0
WTForms==3.0.1
zipp==3.8.1
zope.interface==5.4.0
